+++ extension/chrome/elements/pgp_block_modules/pgp_block_error_module.ts
@@ -2,15 +2,15 @@
 
 'use strict';
 
-import { PgpBlockView } from './pgp_block.js';
-import { Xss } from '../../js/common/platform/xss.js';
-import { Api } from '../../js/common/api/api.js';
-import { BrowserMsg } from '../../js/common/extension.js';
-import { Catch } from '../../js/common/platform/catch.js';
-import { FormatError } from '../../js/common/core/pgp.js';
-import { Lang } from '../../js/common/lang.js';
-import { Store } from '../../js/common/platform/store.js';
-import { Ui } from '../../js/common/browser.js';
+import { PgpBlockView } from '../pgp_block.js';
+import { Xss } from '../../../js/common/platform/xss.js';
+import { Api } from '../../../js/common/api/api.js';
+import { BrowserMsg } from '../../../js/common/extension.js';
+import { Catch } from '../../../js/common/platform/catch.js';
+import { FormatError } from '../../../js/common/core/pgp.js';
+import { Lang } from '../../../js/common/lang.js';
+import { Store } from '../../../js/common/platform/store.js';
+import { Ui } from '../../../js/common/browser.js';
 
 export class PgpBlockViewErrorModule {
