+++ extension/chrome/elements/pgp_block_modules/pgp_block_render_module.ts
@@ -2,16 +2,16 @@
 
 'use strict';
 
-import { PgpBlockView } from './pgp_block.js';
-import { Xss } from '../../js/common/platform/xss.js';
-import { MsgBlock, Mime } from '../../js/common/core/mime.js';
-import { Att } from '../../js/common/core/att.js';
-import { Buf } from '../../js/common/core/buf.js';
-import { VerifyRes, PgpMsg } from '../../js/common/core/pgp.js';
-import { BrowserMsg } from '../../js/common/extension.js';
-import { Ui } from '../../js/common/browser.js';
-import { Store } from '../../js/common/platform/store.js';
-import { Catch } from '../../js/common/platform/catch.js';
+import { PgpBlockView } from '../pgp_block.js';
+import { Xss } from '../../../js/common/platform/xss.js';
+import { MsgBlock, Mime } from '../../../js/common/core/mime.js';
+import { Att } from '../../../js/common/core/att.js';
+import { Buf } from '../../../js/common/core/buf.js';
+import { VerifyRes, PgpMsg } from '../../../js/common/core/pgp.js';
+import { BrowserMsg } from '../../../js/common/extension.js';
+import { Ui } from '../../../js/common/browser.js';
+import { Store } from '../../../js/common/platform/store.js';
+import { Catch } from '../../../js/common/platform/catch.js';
 
 export class PgpBlockViewRenderModule {
